use assert utils instead of require panics
