+++ tests/integration/staking/keeper/keeper_test.go
@@ -52,25 +52,3 @@ func initKeeperFixture(t *testing.T) *keeperFixture {
 
 	return f
 }
-
-func assertNotPanics(t *testing.T, f func()) {
-	defer func() {
-		if r := recover(); r != nil {
-			t.Errorf("should not panic: %v", r)
-		}
-	}()
-	f()
-}
-
-func assertPanics(t *testing.T, f func()) {
-	panicked := false
-	defer func() {
-		if r := recover(); r != nil {
-			panicked = true
-		}
-	}()
-	f()
-	if !panicked {
-		t.Errorf("should panic")
-	}
-}
