+++ tests/integration/staking/keeper/slash_test.go
@@ -9,6 +9,7 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	"gotest.tools/v3/assert"
 
+	sdktestutil "github.com/cosmos/cosmos-sdk/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktestutil "github.com/cosmos/cosmos-sdk/x/bank/testutil"
 	"github.com/cosmos/cosmos-sdk/x/staking/keeper"
@@ -381,7 +382,7 @@ func TestSlashWithRedelegation(t *testing.T) {
 	_, found := app.StakingKeeper.GetValidatorByConsAddr(ctx, consAddr)
 	assert.Assert(t, found)
 
-	assertNotPanics(t, func() {
+	sdktestutil.AssertNotPanics(t, func() {
 		app.StakingKeeper.Slash(ctx, consAddr, 10, 10, fraction)
 	})
 	burnAmount := sdk.NewDecFromInt(app.StakingKeeper.TokensFromConsensusPower(ctx, 10)).Mul(fraction).TruncateInt()
@@ -414,7 +415,7 @@ func TestSlashWithRedelegation(t *testing.T) {
 	_, found = app.StakingKeeper.GetValidatorByConsAddr(ctx, consAddr)
 	assert.Assert(t, found)
 
-	assertNotPanics(t, func() {
+	sdktestutil.AssertNotPanics(t, func() {
 		app.StakingKeeper.Slash(ctx, consAddr, 10, 10, math.LegacyOneDec())
 	})
 	burnAmount = app.StakingKeeper.TokensFromConsensusPower(ctx, 7)
@@ -450,7 +451,7 @@ func TestSlashWithRedelegation(t *testing.T) {
 	_, found = app.StakingKeeper.GetValidatorByConsAddr(ctx, consAddr)
 	assert.Assert(t, found)
 
-	assertNotPanics(t, func() {
+	sdktestutil.AssertNotPanics(t, func() {
 		app.StakingKeeper.Slash(ctx, consAddr, 10, 10, math.LegacyOneDec())
 	})
 
@@ -485,7 +486,7 @@ func TestSlashWithRedelegation(t *testing.T) {
 	validator, _ = app.StakingKeeper.GetValidatorByConsAddr(ctx, consAddr)
 	assert.Equal(t, validator.GetStatus(), types.Unbonding)
 
-	assertNotPanics(t, func() {
+	sdktestutil.AssertNotPanics(t, func() {
 		app.StakingKeeper.Slash(ctx, consAddr, 10, 10, math.LegacyOneDec())
 	})
